+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/TypesafeBackendImpl.kt
@@ -7,6 +7,8 @@ import cash.z.ecc.android.sdk.internal.model.JniSubtreeRoot
 import cash.z.ecc.android.sdk.internal.model.ScanRange
 import cash.z.ecc.android.sdk.internal.model.ScanSummary
 import cash.z.ecc.android.sdk.internal.model.SubtreeRoot
+import cash.z.ecc.android.sdk.internal.model.TransactionDataRequest
+import cash.z.ecc.android.sdk.internal.model.TransactionStatus
 import cash.z.ecc.android.sdk.internal.model.TreeState
 import cash.z.ecc.android.sdk.internal.model.WalletSummary
 import cash.z.ecc.android.sdk.internal.model.ZcashProtocol
@@ -231,6 +233,14 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
         limit: Long
     ): ScanSummary = ScanSummary.new(backend.scanBlocks(fromHeight.value, fromState.encoded, limit), network)
 
+    override suspend fun transactionDataRequests(): List<TransactionDataRequest> =
+        backend.transactionDataRequests().map { jniRequest ->
+            TransactionDataRequest.new(
+                jniRequest,
+                network
+            )
+        }
+
     override suspend fun getWalletSummary(): WalletSummary? =
         backend.getWalletSummary()?.let { jniWalletSummary ->
             WalletSummary.new(jniWalletSummary)
@@ -244,7 +254,19 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
             )
         }
 
-    override suspend fun decryptAndStoreTransaction(tx: ByteArray) = backend.decryptAndStoreTransaction(tx)
+    override suspend fun decryptAndStoreTransaction(
+        tx: ByteArray,
+        minedHeight: BlockHeight?
+    ) = backend
+        .decryptAndStoreTransaction(tx, minedHeight?.value)
+
+    override suspend fun setTransactionStatus(
+        txId: ByteArray,
+        status: TransactionStatus
+    ) = backend.setTransactionStatus(
+        txId = txId,
+        status = status.toPrimitiveValue()
+    )
 
     override fun getSaplingReceiver(ua: String): String? = backend.getSaplingReceiver(ua)
 
@@ -262,4 +284,6 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
     override fun isValidTransparentAddr(addr: String): Boolean = backend.isValidTransparentAddr(addr)
 
     override fun isValidUnifiedAddr(addr: String): Boolean = backend.isValidUnifiedAddr(addr)
+
+    override fun isValidTexAddr(addr: String): Boolean = backend.isValidTexAddr(addr)
 }